// packages/sdk/src/lib/abi/RollupUserLogic.ts
export type ExecutionStateStructOutput = [GlobalStateStructOutput, number] & {
  globalState: GlobalStateStructOutput;
  machineStatus: number;
};

export type AssertionStruct = {
  beforeState: ExecutionStateStruct;
  afterState: ExecutionStateStruct;
  numBlocks: BigNumberish;
};

export type AssertionStructOutput = [
  ExecutionStateStructOutput,
  ExecutionStateStructOutput,
  BigNumber
] & {
  beforeState: ExecutionStateStructOutput;
  afterState: ExecutionStateStructOutput;
  numBlocks: BigNumber;
};
// ...
export type NodeCreatedEvent = TypedEvent<
  [
    BigNumber,
    string,
    string,
    string,
    AssertionStructOutput,
    string,
    string,
    BigNumber
  ],
  {
    nodeNum: BigNumber;
    parentNodeHash: string;
    nodeHash: string;
    executionHash: string;
    assertion: AssertionStructOutput;
    afterInboxBatchAcc: string;
    wasmModuleRoot: string;
    inboxMaxCount: BigNumber;
  }
>;