// packages/sdk/src/lib/abi-bold/BoldRollupUserLogic.ts
export type BeforeStateDataStructOutput = [
  string,
  string,
  ConfigDataStructOutput
] & {
  prevPrevAssertionHash: string
  sequencerBatchAcc: string
  configData: ConfigDataStructOutput
}
// ...
export type AssertionStateStructOutput = [
  GlobalStateStructOutput,
  number,
  string
] & {
  globalState: GlobalStateStructOutput
  machineStatus: number
  endHistoryRoot: string
}

export type AssertionInputsStruct = {
  beforeStateData: BeforeStateDataStruct
  beforeState: AssertionStateStruct
  afterState: AssertionStateStruct
}

export type AssertionInputsStructOutput = [
  BeforeStateDataStructOutput,
  AssertionStateStructOutput,
  AssertionStateStructOutput
] & {
  beforeStateData: BeforeStateDataStructOutput
  beforeState: AssertionStateStructOutput
  afterState: AssertionStateStructOutput
}
// ...
export type AssertionCreatedEvent = TypedEvent<
  [
    string,
    string,
    AssertionInputsStructOutput,
    string,
    BigNumber,
    string,
    BigNumber,
    string,
    BigNumber
  ],
  {
    assertionHash: string
    parentAssertionHash: string
    assertion: AssertionInputsStructOutput
    afterInboxBatchAcc: string
    inboxMaxCount: BigNumber
    wasmModuleRoot: string
    requiredStake: BigNumber
    challengeManager: string
    confirmPeriodBlocks: BigNumber
  }
>